pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["devriez <a.moiseikin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
dj-database-url = "^0.5.0"
psycopg2 = "^2.9.3"
django-bootstrap4 = "^22.1"
django-filter = "^21.1"
rollbar = "^0.16.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
